+++ lib/JSExtract.py
@@ -1,14 +1,15 @@
+from termcolor import colored
 from traceback import print_exc
 from re import search, IGNORECASE
-from termcolor import colored
 from urllib.parse import urlparse
 
 from lib.Engine import Engine
-from lib.Globals import dom_sources_regex, dom_sinks_regex
-from lib.Globals import url_regex, subdomain_regex
-from lib.Globals import ColorObj
-from lib.Functions import manage_output
 from lib.PathFunctions import PathFunction
+from lib.Globals import ColorObj, base64char, hexchar
+from lib.Globals import dom_sources_regex, dom_sinks_regex
+from lib.Globals import url_regex, subdomain_regex, path_regex
+from lib.Globals import single_path_regex
+from lib.Functions import manage_output, shannon_entropy
 
 JSE = Engine()
 FPathApp = PathFunction()
@@ -37,41 +38,101 @@ class JSExtract:
                     output_list.append(manage_output(f"{jscomment.strip(' ')} <--- Comments\n"))
             for line in jstext:
                 line = line.strip(' ').rstrip('{').rstrip(' ').lstrip('}').lstrip(' ')
-                for dom_source in dom_sources_regex:
-                    if search(dom_source, line, IGNORECASE):
-                        print(f"{ColorObj.good} Found Dom XSS Source: {colored(line.strip(' '), color='red', attrs=['bold'])}")
-                        output_list.append(manage_output(f"{line.strip(' ')} <--- DomXSS Source {dom_source}\n", color=dom_source))
-                        self.continuer = 1
-                        break
-                if self.continuer:
-                    self.continuer = 0
-                    continue
-                for dom_sink in dom_sinks_regex:
-                    if search(dom_sink, line, IGNORECASE):
-                        print(f"{ColorObj.good} Found Dom XSS Sink: {colored(line.strip(' '), color='red', attrs=['bold'])}")
-                        output_list.append(manage_output(f"{line.strip(' ')} <--- DomXSS Sink {dom_sink}\n", color=dom_sink))
-                        self.continuer = 1
-                        break
-                if self.continuer:
-                    self.continuer = 0
-                    continue
-                if search(url_regex, line):
-                    print(f"{ColorObj.good} Found endpoint: {colored(line.strip(' '), color='red', attrs=['bold'])}")
-                    output_list.append(manage_output(f"{line.strip(' ')} <--- Endpoint \n"))
-                    continue
-                if self.argv.domain:
-                    subdomain = subdomain_regex(self.argv.domain)
-                    if search(subdomain, line, IGNORECASE):
-                        actual_sub = [word for word in line.split(' ') if search(subdomain, word, IGNORECASE)][0].replace(';', '').strip('"').strip("'")
-                        print(f"{ColorObj.good} Found subdomain: {colored(actual_sub, color='red', attrs=['bold'])}")
-                        output_list.append(manage_output(f"{actual_sub} <--- SubRegex {subdomain}\n", color=actual_sub))
-                        continue
-                if self.argv.enable_entropy:
-                    for word in line.split(' '):
-                        if len(word) > 5:
-                            if float(shannon_entropy(word, base64char)) > float(3.5) or float(shannon_entropy(word, hexchar)) > float(3.6):
-                                print(f"{ColorObj.good} Suspicious data: {colored(word, color='red', attrs=['bold'])}")
-                                output_list.append(manage_output(f"{word} <--- Entropy \n"))
+                output_list.append(self.domsource_extract(line))
+                if self.continuer: self.continuer = 0; continue
+                output_list.append(self.domsink_extract(line))
+                if self.continuer: self.continuer = 0; continue
+                output_list.append(self.url_extract(line))
+                if self.continuer: self.continuer = 0; continue
+                output_list.append(self.path_extract(line))
+                if self.continuer: self.continuer = 0; continue
+                output_list.append(self.subdomain_extract(line))
+                if self.continuer: self.continuer = 0; continue
+                output_list.append(self.shannon_extract(line))
+                if self.continuer: self.continuer = 0; continue
             return output_list
         except Exception:
             print_exc()
+
+    def domsource_extract(self, line):
+        output_list = ""
+        for dom_source in dom_sources_regex:
+            if search(dom_source, line, IGNORECASE):
+                print(f"{ColorObj.good} Found Dom XSS Source: {colored(line.strip(' '), color='red', attrs=['bold'])}")
+                output_list = manage_output(f"{line.strip(' ')} <--- DomXSS Source {dom_source}\n", color=dom_source)
+                self.continuer = 1
+                return output_list
+        return ""
+
+    def domsink_extract(self, line):
+        output_list = ""
+        for dom_source in dom_sinks_regex:
+            if search(dom_source, line, IGNORECASE):
+                print(f"{ColorObj.good} Found Dom XSS Source: {colored(line.strip(' '), color='red', attrs=['bold'])}")
+                output_list = manage_output(f"{line.strip(' ')} <--- DomXSS Source {dom_source}\n", color=dom_source)
+                self.continuer = 1
+                return output_list
+        return ""
+
+    def subdomain_extract(self, line):
+        output_list = ""
+        if not self.argv.domain:
+            return ""
+        subdomain = subdomain_regex(self.argv.domain)
+        if search(subdomain, line, IGNORECASE):
+            sub = [w for w in line.split(' ') if search(subdomain, w, IGNORECASE)][0].replace(';', '').strip('"').strip("'")
+            print(f"{ColorObj.good} Found subdomain: {colored(sub, color='red', attrs=['bold'])}")
+            output_list = manage_output(f"{sub} <--- SubRegex {subdomain}\n", color=sub)
+            self.continuer = 1
+            return output_list
+        return ""
+    
+    def url_extract(self, line):
+        output_list = ""
+        if search(url_regex, line):
+            line = search(url_regex, line).group()
+            print(f"{ColorObj.good} Found endpoint: {colored(line.strip(' '), color='red', attrs=['bold'])}")
+            output_list = manage_output(f"{line.strip(' ')} <--- Endpoint \n")
+            self.continuer = 1
+            return output_list
+        return ""
+    
+    def path_extract(self, line):
+        output_list = ""
+        if search(path_regex, line):
+            line = search(path_regex, line).group()
+            print(f"{ColorObj.good} Found endpoint: {colored(line.strip(' '), color='red', attrs=['bold'])}")
+            output_list = manage_output(f"{line.strip(' ')} <--- Endpoint \n")
+            self.continuer = 1
+        elif search(single_path_regex, line):
+            line = self.reduce_string(search(single_path_regex, line).group(), args=['"', "'"])
+            print(f"{ColorObj.good} Found endpoint: {colored(line.strip(' '), color='red', attrs=['bold'])}")
+            output_list = manage_output(f"{line.strip(' ')} <--- Endpoint \n")
+            self.continuer = 1
+            return output_list
+        return "" 
+
+    def shannon_extract(self, line):
+        output_list = ""
+        if self.argv.enable_entropy:
+            for word in line.split(' '):
+                if len(word) > 5:
+                    if float(shannon_entropy(word, base64char)) > float(3.43) or float(shannon_entropy(word, hexchar)) > float(3.5):
+                        word = self.reduce_string(word.rstrip(';'), args=['"', "'"])
+                        print(f"{ColorObj.good} Suspicious data: {colored(word, color='red', attrs=['bold'])}")
+                        output_list = manage_output(f"{word} <--- Entropy \n")
+                        self.continuer = 1
+                        return output_list
+        return ""
+    
+    def reduce_string(self, line, args):
+        line = line.rstrip('//').rstrip(';')
+        for arg in args:
+            if arg in line and line[0] == arg and line[-1] == arg:
+                line = line[1:-1]
+                return line
+            elif arg == "(" or arg == ")":
+                if line[0] == "(" and line[-1] == ")":
+                    line = line[1:-1]
+                    return line
+        return line